progress on team creation page
