+++ Assets/lang/en.json
@@ -489,25 +489,28 @@
                 "teams": "TEAMS"
             }
         },
-        "register": {
-            "title": "REGISTER YOUR TEAM",
+        "create": {
+            "title": "CREATE A TEAM",
             "info1": "Registrations will end on",
             "info2": "Please ensure all information is correct before finalizing.",
             "teamName": "TEAM NAME",
-            "maxCharacters": "max 20 characters",
-            "teamBanner": "TEAM BANNER",
+            "teamLength": "5 - 20 characters",
+            "teamAcronym": "TEAM ACRONYM",
+            "acronymLength": "2 - 4 characters",
+            "teamAvatar": "TEAM AVATAR",
             "upload": "UPLOAD",
             "fileSizeLimit": "max 5mb limit. accepted file types: .jpg .png",
             "fileSizeError": "file size too big! please upload a smaller image",
+            "fileTypeError": "incorrect file type! please upload a .jpg or .png",
             "preview": "PREVIEW",
-            "bannersInfo1": "Captains can modify team names and banners before registrations close.",
-            "bannersInfo2": "All names and banners have to be approved by staff.",
-            "bannersInfo3": "If deemed inappropriate, you will be notified to make changes.",
+            "avatarInfo1": "Captains can modify team names and avatars before registrations close.",
+            "avatarInfo2": "All names and avatars will be approved by staff.",
+            "avatarInfo3": "If deemed inappropriate, you will be notified to make changes.",
             "teamManagement": "TEAM MANAGEMENT",
-            "teamCaptains": "Team <span>captains</span> are able to play in matches but occupy one player slot.",
-            "teamManagers": "<div>Team <span>managers</span> cannot play in matches but do not occupy player slots, therefore</div><div>a team can have one manager and up to 6 players.</div>",
-            "selectManager": "Select this checkbox if you are acting as the team <span>manager</span>.",
-            "confirmManager": "<div>I am a manager for this team and I understand I will not be able to participate</div> <div>any matches.</div>",
+            "teamManagers": "<div>Team <span>managers</span> (that is you) have the option to play and be captain simultaneously</div> <div><span>OR</span> not play with the team and only be captain.</div>",
+            "teamManagersFree": "If you wish to not play, and only act as a manager, you will <span>not</span> take up a player slot.",
+            "selectManager": "Select this checkbox if you are <span>not</span> going to play with the team, and only act as <span>manager</span>.",
+            "confirmManager": "<div>By checking this, I will only act as manager, and I understand I will not be able to participate</div> <div>in any matches.</div>",
             "continue": "CONTINUE TO NEXT STEP"
         }
     },
